Updated tests to use pytest for cleaner outputs and code.
